3_4_captcha/requirements.txt
pandas
scikit-learn
fastapi
uvicorn
python-multipart
torch
torchvision
Pillow
numpy
